# Script/Kaelthas: Telonicus, timing, and misc. fixes