--- a/sql/updates/world/059_kaelthas_polishing.sql
+++ b/sql/updates/world/059_kaelthas_polishing.sql
@@ -1,0 +1,2 @@
+-- Kael'thas Advisor Thaladred the Darkener Gaze Emote Update
+UPDATE `script_texts` SET `content_default`='sets his eyes on $N!' WHERE `entry` = -1550037;